requirements.txt
python-dateutil
SpeechRecognition
pyttsx3
requests
pytest
flask
pyaudio